+++ Serveur/src/prismaClient.ts
@@ -1,0 +1,11 @@
+import { PrismaClient } from '@prisma/client';
+
+let prisma: PrismaClient;
+
+if (!global.prisma) {
+    global.prisma = new PrismaClient();
+}
+
+prisma = global.prisma;
+
+export default prisma;
