+++ Serveur/tsconfig.json
@@ -8,5 +8,5 @@
     "moduleResolution": "node",
     "esModuleInterop": true,
     "types": ["node"]
-  }
+  },
 }
